# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "description": "The request body does not accept a password. The API generates a temporary password, emails it to the user, marks mustChangePassword as true, and returns only safe user fields.",
        "operationId": "UsersController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully. Temporary password is sent by email and is not returned in the response.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 201,
                    "message": "User created successfully",
                    "data": {
                      "id": "user-id",
                      "email": "newuser@example.com",
                      "username": "newuser",
                      "isActive": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid role IDs or validation failed"
          },
          "409": {
            "description": "User already exists"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Create a user with a generated temporary password sent by email",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_getUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Users retrieved successfully",
                    "data": {
                      "items": [
                        {
                          "id": "user-1",
                          "email": "user1@example.com",
                          "username": "user1",
                          "isActive": true
                        },
                        {
                          "id": "user-2",
                          "email": "user2@example.com",
                          "username": "user2",
                          "isActive": true
                        }
                      ],
                      "total": 2,
                      "page": 1,
                      "limit": 10,
                      "totalPages": 1
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Retrieve users with pagination and search",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "User retrieved successfully",
                    "data": {
                      "id": "user-id",
                      "email": "user@example.com",
                      "username": "user",
                      "isActive": true,
                      "roles": []
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Retrieve a single user by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "User updated successfully",
                    "data": {
                      "id": "user-id",
                      "email": "user@example.com",
                      "username": "user",
                      "firstName": "John",
                      "lastName": "Doe",
                      "isActive": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Update safe user profile and status fields",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "description": "This endpoint deactivates the account instead of removing the database row. A user cannot delete their own account.",
        "operationId": "UsersController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "User deleted successfully",
                    "data": {
                      "id": "user-id",
                      "email": "user@example.com",
                      "username": "user",
                      "isActive": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cannot delete your own account"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Soft-delete a user by deactivating the account and revoking sessions",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/reset-password": {
      "post": {
        "description": "Admin reset flow. Generates a new temporary password, emails it to the selected user, sets mustChangePassword to true, and revokes active sessions and refresh tokens.",
        "operationId": "UsersController_resetUserPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Temporary password emailed. The generated password is not returned in the response.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Temporary password generated and emailed successfully",
                    "data": {
                      "id": "user-id",
                      "email": "user@example.com",
                      "username": "user",
                      "isActive": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Generate a temporary password for a user and require password change on next login",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "SaaS Multi-Tenant API",
    "description": "Auto-generated OpenAPI specification for the SaaS Demo application.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token to authenticate",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "new.user@example.com"
          },
          "username": {
            "type": "string",
            "example": "newuser"
          },
          "firstName": {
            "type": "string",
            "example": "New"
          },
          "lastName": {
            "type": "string",
            "example": "User"
          },
          "roleIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "email"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "avatarUrl": {
            "type": "string",
            "example": "https://example.com/avatar.png"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  }
}